Audit Log Modules and Hive Hook refactoring
PURPOSE

Add # of mappers/reducers per stage to the logged hive metrics (MySQL), as well total cpu time spent, and various counters that hive keeps
Refactor the code to be compatible with the public reair repo (i.e. make the internal reair repo a fork of the public one) and merge in the
OTHER CHANGES

Refactored the hooks class to improve modularization (and moved the environment variable hooks into it's own table).
AuditLogHook became AuditLogHook which depends on ObjectsLogHook + MapRedStatsLogHook + EnvVarsLogHook (+ the mapreduce stats loghook I add
The new files prefixed with DataInfra contain the airbnb-specific features of their open-source counterparts, e.g. DataInfraAuditLogHookTes
Added section in pom.xml from the current internal fork in data-infra to allow for generating all-hook jars.
Removed some internal-only code that was left in the open-source tests.
Start introducing Jackson for JSON parsing as this is the standard.
Replaced the ArrayList .add .add ... pattern in tests with Lists.newArrayList.